Move kraken related code to epynomous file
